gradio_app/requirements_gradio.txt
gradio
google-cloud-aiplatform
pandas
numpy
requests